Moved consts & utils to folder "utils"
